build(tesseratos): make it an application
