+++ docs/Doxyfile
@@ -10,8 +10,7 @@ INPUT = @DOXYGEN_PROJECT_ROOT@/docs/pages \
         @DOXYGEN_PROJECT_ROOT@/core/samples \
         @DOXYGEN_PROJECT_ROOT@/engine/include \
         @DOXYGEN_PROJECT_ROOT@/engine/samples \
-        @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/include \
-        @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/samples \
+        @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/src \
         @DOXYGEN_PROJECT_ROOT@/api/core/include \
         @DOXYGEN_PROJECT_ROOT@/api/core/samples \
         @DOXYGEN_PROJECT_ROOT@/api/engine/include \
@@ -19,7 +18,6 @@ INPUT = @DOXYGEN_PROJECT_ROOT@/docs/pages \
 
 EXAMPLE_PATH = @DOXYGEN_PROJECT_ROOT@/core/samples \
                @DOXYGEN_PROJECT_ROOT@/engine/samples \
-               @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/samples \
                @DOXYGEN_PROJECT_ROOT@/api/core/samples \
                @DOXYGEN_PROJECT_ROOT@/api/engine/samples
 EXTRACT_ALL = YES
@@ -30,7 +28,6 @@ OUTPUT_DIRECTORY = @DOXYGEN_OUTPUT_DIRECTORY@
 IMAGE_PATH = @DOXYGEN_PROJECT_ROOT@/docs/images \
              @DOXYGEN_PROJECT_ROOT@/core/samples \
              @DOXYGEN_PROJECT_ROOT@/engine/samples \
-             @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/samples \
              @DOXYGEN_PROJECT_ROOT@/api/core/samples \
              @DOXYGEN_PROJECT_ROOT@/api/engine/samples
 
@@ -44,8 +41,7 @@ STRIP_FROM_PATH = @DOXYGEN_PROJECT_ROOT@/core/include/cubos \
                   @DOXYGEN_PROJECT_ROOT@/core/samples \
                   @DOXYGEN_PROJECT_ROOT@/engine/include/cubos \
                   @DOXYGEN_PROJECT_ROOT@/engine/samples \
-                  @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/include \
-                  @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/samples \
+                  @DOXYGEN_PROJECT_ROOT@/tools/tesseratos/src \
                   @DOXYGEN_PROJECT_ROOT@/api/core/include \
                   @DOXYGEN_PROJECT_ROOT@/api/core/samples \
                   @DOXYGEN_PROJECT_ROOT@/api/engine/include \
