+++ tools/tesseratos/assets/scenes/main.cubos
@@ -1,0 +1,3 @@
+{
+    "entities": {}
+}
