+++ tools/tesseratos/assets/scenes/main.cubos.meta
@@ -1,0 +1,3 @@
+{
+    "id": "b0102b7d-8c55-45d8-a848-5c4b2c8eac70"
+}
