+++ tools/tesseratos/assets/tesseratos.bind
