+++ tools/tesseratos/assets/tesseratos.bind.meta
