--- tools/tesseratos/samples/main/assets/scenes/main.cubos
@@ -1,134 +1,0 @@
-{
-  "entities": {
-    "child": {
-      "cubos::engine::Position": {
-        "x": 3.0,
-        "y": 3.0,
-        "z": 3.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "child1": {
-      "cubos::engine::Position": {
-        "x": 4.0,
-        "y": 4.0,
-        "z": 4.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "child2": {
-      "cubos::engine::Position": {
-        "x": 5.0,
-        "y": 5.0,
-        "z": 5.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "child3": {
-      "cubos::engine::Position": {
-        "x": 7.0,
-        "y": 7.0,
-        "z": 7.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "child4": {
-      "cubos::engine::Position": {
-        "x": 9.0,
-        "y": 9.0,
-        "z": 9.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "child5": {
-      "cubos::engine::Position": {
-        "x": 10.0,
-        "y": 10.0,
-        "z": 10.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "main": {
-      "cubos::engine::Position": {
-        "x": 1.0,
-        "y": 1.0,
-        "z": 1.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "parent": {},
-    "parent1": {
-      "cubos::engine::Position": {
-        "x": 6.0,
-        "y": 6.0,
-        "z": 6.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    },
-    "parent2": {
-      "cubos::engine::Position": {
-        "x": 8.0,
-        "y": 8.0,
-        "z": 8.0
-      },
-      "cubos::engine::Rotation": {
-        "w": 1.0,
-        "x": 0.0,
-        "y": 0.0,
-        "z": 0.0
-      },
-      "cubos::engine::Scale": 1.0
-    }
-  },
-  "imports": {
-    "sub1": "00d86ba8-5f34-440f-a180-d9d12c8e8b91"
-  }
-}
