--- tools/tesseratos/samples/main/assets/scenes/main.cubos.meta
@@ -1,3 +1,0 @@
-{
-    "id": "10d86ba8-5f34-440f-a180-d9d12c8e8b91"
-}
