--- tools/tesseratos/samples/main/assets/scenes/sub.cubos
@@ -1,14 +1,0 @@
-{
-  "entities": {
-    "subsub.main": {
-      "cubos::engine::Position": {
-        "x": 4.0,
-        "y": 3.0,
-        "z": 3.0
-      }
-    }
-  },
-  "imports": {
-    "subsub": "3e17fb64-8abf-4c9a-8392-46fb6969e5e2"
-  }
-}
