--- tools/tesseratos/samples/main/assets/scenes/sub.cubos.meta
@@ -1,4 +1,0 @@
-{
-    "id": "00d86ba8-5f34-440f-a180-d9d12c8e8b91",
-    "path": "/assets/scenes/sub.cubos"
-}
