--- tools/tesseratos/samples/main/assets/scenes/subsub.cubos.meta
@@ -1,4 +1,0 @@
-{
-    "id": "3e17fb64-8abf-4c9a-8392-46fb6969e5e2",
-    "path": "/assets/scenes/subsub.cubos"
-}
