--- tools/tesseratos/samples/main/assets/voxels/car.grd.meta
@@ -1,3 +1,0 @@
-{
-    "id": "059c16e7-a439-44c7-9bdc-6e069dba0c75"
-}
