--- tools/tesseratos/samples/main/assets/voxels/main.pal.meta
@@ -1,3 +1,0 @@
-{
-    "id": "1aa5e234-28cb-4386-99b4-39386b0fc215"
-}
