+++ tools/tesseratos/src/tesseratos/asset_explorer/popup.cpp
@@ -1,8 +1,8 @@
+#include "popup.hpp"
+
 #include <imgui.h>
 
 #include <cubos/core/reflection/type.hpp>
-
-#include <tesseratos/asset_explorer/popup.hpp>
 
 using cubos::core::reflection::Type;
