+++ tools/tesseratos/src/tesseratos/debug_camera/plugin.hpp
