+++ tools/tesseratos/src/tesseratos/voxel_palette_editor/plugin.hpp
